perf: decompose SidePanel React island into Astro + small islands (#227)
Replace the 1134-line monolithic client:load island with a static Astro
shell and focused reactive islands so the sidebar renders in the initial
HTML instead of after JS hydration.
